break overlay into separate component
